[#31] Modify file directory, set rightConfig props
