+++ src/components/Nav/index.tsx
@@ -1,16 +1,18 @@
 import Link from "next/link";
+import { ReactElement } from "react";
 import styled from "styled-components";
-import AddIcon from "@mui/icons-material/Add";
 
-function Nav() {
+interface INavProps {
+  rightConfig?: ReactElement;
+}
+
+function Nav({ rightConfig }: INavProps) {
   return (
     <StyledNavWrapper>
       <Link href="/Main">
         <StyledLogo>DontWorry</StyledLogo>
       </Link>
-      <Link href="/Main/Write">
-        <StyledWritePageBtn />
-      </Link>
+      {rightConfig && <>{rightConfig}</>}
     </StyledNavWrapper>
   );
 }
@@ -33,9 +35,5 @@ const StyledNavWrapper = styled.nav`
 `;
 
 const StyledLogo = styled.h1`
-  cursor: pointer;
-`;
-
-const StyledWritePageBtn = styled(AddIcon)`
   cursor: pointer;
 `;
